pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fraudlens"
version = "1.0.0"
description = "Multimodal fraud detection pipeline combining computer vision (SigLIP 2), NLP (DistilBERT), and structured transaction analysis with learned attention fusion."
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.10"
authors = [
    {name = "Harshal Patel"},
]

dependencies = [
    # Core ML
    "torch>=2.1.0",
    "torchvision>=0.16.0",
    "transformers>=4.40.0",
    "scikit-learn>=1.4.0",
    # Data
    "pandas>=2.1.0",
    "numpy>=1.26.0",
    "Pillow>=10.0.0",
    # Explainability
    "shap>=0.44.0",
    "captum>=0.7.0",
    # Serving
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.5.0",
    "python-multipart>=0.0.9",
    # Config
    "pyyaml>=6.0",
    # Logging
    "tensorboard>=2.15.0",
    "tqdm>=4.66.0",
    # Utilities
    "kaggle>=1.6.0",
    # Visualization
    "matplotlib>=3.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "httpx>=0.27.0",
    "ruff>=0.3.0",
    "mypy>=1.8.0",
]

[tool.setuptools.packages.find]
where = ["."]
include = ["src*"]

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true
